requirement.txt
openai
openai-whisper
yt-dlp
ffmpeg(whisper, yt-dlp)